테스트 관련 파일 및 코드정리 및 위치 변경
